Implemented showing recipe with details

+++ KulinarkaBackend/Models/RecipeIngredient.cs
@@ -22,13 +22,13 @@ namespace Kulinarka.Models
         public int Id { get; set; }
         public int RecipeId { get; set; }
         public int IngredientId { get; set; }
-        public float Amount { get; set; }
+        public double Amount { get; set; }
         public MeasurementUnit MeasurementUnit { get; set; }
         [JsonIgnore]
         public virtual Recipe? Recipe { get; set; }
         [JsonIgnore]
         public virtual Ingredient? Ingredient { get; set; }
-        public RecipeIngredient(int recipeId, int ingredientId, float amount, MeasurementUnit measurementUnit)
+        public RecipeIngredient(int recipeId, int ingredientId, double amount, MeasurementUnit measurementUnit)
         {
             RecipeId = recipeId;
             IngredientId = ingredientId;
